Revert "[asset graph] Minor sidebar refactor + Fix asset graph sidebar tree view auto-scrolling (#32006)"
This reverts commit 49e9d12dfdf2f1b02429783a6c03062a787019f3.
