+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/AssetGraphExplorer.tsx
@@ -699,8 +699,7 @@ const AssetGraphExplorerWithData = ({
   if (!nextLayoutLoading && isInitialLayout.current) {
     isInitialLayout.current = false;
   }
-  const sidebarLoading = dataLoading && isInitialLayout.current;
-  const loading = sidebarLoading || nextLayoutLoading;
+  const loading = (layoutLoading || dataLoading) && isInitialLayout.current;
 
   const [errorState, setErrorState] = useState<SyntaxError[]>([]);
 
@@ -735,7 +734,7 @@ const AssetGraphExplorerWithData = ({
       firstMinSize={400}
       secondMinSize={400}
       first={
-        sidebarLoading ? (
+        loading ? (
           <LoadingContainer>
             <Box margin={{bottom: 24}}>Loading assets…</Box>
             <Spinner purpose="page" />
