+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/sidebar/AssetSidebarListView.tsx
@@ -22,11 +22,7 @@ export interface AssetSidebarListViewProps {
   setSelectedNode: React.Dispatch<
     React.SetStateAction<null | {id: string; path: string} | {id: string}>
   >;
-  selectNode: (
-    e: React.MouseEvent<any> | React.KeyboardEvent<any>,
-    nodeId: string,
-    sidebarNodeInfo?: {path: string; direction: 'root-to-leaf' | 'leaf-to-root'},
-  ) => void;
+  selectNode: (e: React.MouseEvent<any> | React.KeyboardEvent<any>, nodeId: string) => void;
   explorerPath: ExplorerPath;
   onChangeExplorerPath: (path: ExplorerPath, mode: 'replace' | 'push') => void;
   onFilterToGroup: (group: AssetGroup) => void;
@@ -125,11 +121,7 @@ export const AssetSidebarListView = ({
             // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
             renderedNodes[(nextIndex + renderedNodes.length) % renderedNodes.length]!;
           setSelectedNode(nextNode);
-          selectNode(
-            e,
-            nextNode.id,
-            'path' in nextNode ? {path: nextNode.path, direction} : undefined,
-          );
+          selectNode(e, nextNode.id);
         } else if (e.code === 'ArrowLeft' || e.code === 'ArrowRight') {
           const open = e.code === 'ArrowRight';
           const node = renderedNodes[indexOfLastSelectedNode];
@@ -181,16 +173,12 @@ export const AssetSidebarListView = ({
                     });
                   }}
                   collapseAllNodes={collapseAllNodes}
-                  selectNode={(e, id, path) => {
-                    selectNode(e, id, path);
+                  selectNode={(e, id) => {
+                    selectNode(e, id);
                   }}
                   selectThisNode={(e) => {
                     setSelectedNode(node);
-                    selectNode(
-                      e,
-                      node.id,
-                      'path' in node ? {path: node.path, direction} : undefined,
-                    );
+                    selectNode(e, node.id);
                   }}
                   explorerPath={explorerPath}
                   onChangeExplorerPath={onChangeExplorerPath}
