+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/sidebar/AssetSidebarNode.tsx
@@ -33,11 +33,7 @@ type AssetSidebarNodeProps = {
   level: number;
   toggleOpen: () => void;
   selectThisNode: (e: React.MouseEvent<any> | React.KeyboardEvent<any>) => void;
-  selectNode: (
-    e: React.MouseEvent<any> | React.KeyboardEvent<any>,
-    nodeId: string,
-    sidebarNodeInfo?: {path: string; direction: 'root-to-leaf' | 'leaf-to-root'},
-  ) => void;
+  selectNode: (e: React.MouseEvent<any> | React.KeyboardEvent<any>, nodeId: string) => void;
   isOpen: boolean;
   isLastSelected: boolean;
   isSelected: boolean;
@@ -133,11 +129,7 @@ export const AssetSidebarNode = (props: AssetSidebarNodeProps) => {
 type AssetSidebarAssetLabelProps = {
   fullAssetGraphData?: GraphData;
   node: AssetNodeMenuProps['node'] & StatusDotNode;
-  selectNode: (
-    e: React.MouseEvent<any> | React.KeyboardEvent<any>,
-    nodeId: string,
-    sidebarNodeInfo?: {path: string; direction: 'root-to-leaf' | 'leaf-to-root'},
-  ) => void;
+  selectNode: (e: React.MouseEvent<any> | React.KeyboardEvent<any>, nodeId: string) => void;
   isLastSelected: boolean;
   isSelected: boolean;
   explorerPath: ExplorerPath;
