--- js_modules/dagster-ui/packages/ui-core/src/asset-graph/sidebar/SidebarHeader.module.css
@@ -1,12 +1,0 @@
-.buttonGroupWrapper > * {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-}
-
-.buttonGroupWrapper > * > * {
-  place-content: center;
-}
-
-.buttonGroupWrapper > * span {
-  flex: initial;
-}
